// schema.graphql/API.ts
export enum BusinessBillingType {
  CPC = "CPC",
  CPM = "CPM",
  MONTHLY_RETAINER = "MONTHLY_RETAINER",
}
// ...
export type OnCreateBusinessSubscriptionVariables = {
  filter?: ModelSubscriptionBusinessFilterInput | null,
  owner?: string | null,
};

export type OnCreateBusinessSubscription = {
  onCreateBusiness?:  {
    __typename: "Business",
    address: string,
    billingType?: BusinessBillingType | null,
    businessType?: BusinessBusinessType | null,
    city: string,
    cpmRate?: number | null,
    createdAt: string,
    currentSpend?: number | null,
    description?: string | null,
    email: string,
    id: string,
    isActive?: boolean | null,
    isVerified?: boolean | null,
    latitude?: number | null,
    logoKey?: string | null,
    logoUrl?: string | null,
    longitude?: number | null,
    monthlyBudget?: number | null,
    name: string,
    owner: string,
    phone?: string | null,
    rejectionReason?: string | null,
    rewards?:  {
      __typename: "ModelRewardConnection",
      nextToken?: string | null,
    } | null,
    serviceRadius?: number | null,
    state: string,
    updatedAt: string,
    verificationStatus?: BusinessVerificationStatus | null,
    website?: string | null,
    zipCode: string,
  } | null,
};

export type OnCreateBusinessAnalyticsSubscriptionVariables = {
  filter?: ModelSubscriptionBusinessAnalyticsFilterInput | null,
  owner?: string | null,
};

export type OnCreateBusinessAnalyticsSubscription = {
  onCreateBusinessAnalytics?:  {
    __typename: "BusinessAnalytics",
    avgChildAge?: number | null,
    businessId: string,
    clicks?: number | null,
    createdAt: string,
    date: string,
    id: string,
    impressions?: number | null,
    owner?: string | null,
    redemptions?: number | null,
    revenue?: number | null,
    spend?: number | null,
    topRewardId?: string | null,
    updatedAt: string,
  } | null,
};

export type OnCreateRewardSubscriptionVariables = {
  filter?: ModelSubscriptionRewardFilterInput | null,
  owner?: string | null,
};

export type OnCreateRewardSubscription = {
  onCreateReward?:  {
    __typename: "Reward",
    availableRadius?: number | null,
    business?:  {
      __typename: "Business",
      address: string,
      billingType?: BusinessBillingType | null,
      businessType?: BusinessBusinessType | null,
      city: string,
      cpmRate?: number | null,
      createdAt: string,
      currentSpend?: number | null,
      description?: string | null,
      email: string,
      id: string,
      isActive?: boolean | null,
      isVerified?: boolean | null,
      latitude?: number | null,
      logoKey?: string | null,
      logoUrl?: string | null,
      longitude?: number | null,
      monthlyBudget?: number | null,
      name: string,
      owner: string,
      phone?: string | null,
      rejectionReason?: string | null,
      serviceRadius?: number | null,
      state: string,
      updatedAt: string,
      verificationStatus?: BusinessVerificationStatus | null,
      website?: string | null,
      zipCode: string,
    } | null,
    businessId: string,
    category?: RewardCategory | null,
    clicks?: number | null,
    createdAt: string,
    currentRedemptions?: number | null,
    dailyLimit?: number | null,
    daysOfWeek?: Array< string | null > | null,
    description: string,
    endDate?: string | null,
    id: string,
    imageKey?: string | null,
    imageUrl?: string | null,
    impressions?: number | null,
    isActive?: boolean | null,
    owner?: string | null,
    pointCost: number,
    redemptionInstructions?: string | null,
    redemptionRecords?:  {
      __typename: "ModelRewardRedemptionConnection",
      nextToken?: string | null,
    } | null,
    redemptionType?: RewardRedemptionType | null,
    redemptions?: number | null,
    retailValue?: number | null,
    rewardType?: RewardRewardType | null,
    startDate: string,
    status?: RewardStatus | null,
    targetAgeRanges?: Array< string | null > | null,
    targetFamilySizes?: Array< number | null > | null,
    termsAndConditions?: string | null,
    title: string,
    totalAvailable?: number | null,
    updatedAt: string,
  } | null,
};

export type OnCreateRewardRedemptionSubscriptionVariables = {
  filter?: ModelSubscriptionRewardRedemptionFilterInput | null,
  owner?: string | null,
};

export type OnCreateRewardRedemptionSubscription = {
  onCreateRewardRedemption?:  {
    __typename: "RewardRedemption",
    businessId?: string | null,
    businessRevenue?: number | null,
    childAge?: number | null,
    childId: string,
    createdAt: string,
    expiresAt?: string | null,
    familyId: string,
    fulfilledAt?: string | null,
    id: string,
    notes?: string | null,
    owner?: string | null,
    pointsSpent: number,
    redemptionCity?: string | null,
    redemptionCode?: string | null,
    redemptionDate?: string | null,
    redemptionStatus?: RewardRedemptionRedemptionStatus | null,
    redemptionZip?: string | null,
    reward?:  {
      __typename: "Reward",
      availableRadius?: number | null,
      businessId: string,
      category?: RewardCategory | null,
      clicks?: number | null,
      createdAt: string,
      currentRedemptions?: number | null,
      dailyLimit?: number | null,
      daysOfWeek?: Array< string | null > | null,
      description: string,
      endDate?: string | null,
      id: string,
      imageKey?: string | null,
      imageUrl?: string | null,
      impressions?: number | null,
      isActive?: boolean | null,
      owner?: string | null,
      pointCost: number,
      redemptionInstructions?: string | null,
      redemptionType?: RewardRedemptionType | null,
      redemptions?: number | null,
      retailValue?: number | null,
      rewardType?: RewardRewardType | null,
      startDate: string,
      status?: RewardStatus | null,
      targetAgeRanges?: Array< string | null > | null,
      targetFamilySizes?: Array< number | null > | null,
      termsAndConditions?: string | null,
      title: string,
      totalAvailable?: number | null,
      updatedAt: string,
    } | null,
    rewardId: string,
    updatedAt: string,
  } | null,
};

export type OnDeleteBusinessSubscriptionVariables = {
  filter?: ModelSubscriptionBusinessFilterInput | null,
  owner?: string | null,
};

export type OnDeleteBusinessSubscription = {
  onDeleteBusiness?:  {
    __typename: "Business",
    address: string,
    billingType?: BusinessBillingType | null,
    businessType?: BusinessBusinessType | null,
    city: string,
    cpmRate?: number | null,
    createdAt: string,
    currentSpend?: number | null,
    description?: string | null,
    email: string,
    id: string,
    isActive?: boolean | null,
    isVerified?: boolean | null,
    latitude?: number | null,
    logoKey?: string | null,
    logoUrl?: string | null,
    longitude?: number | null,
    monthlyBudget?: number | null,
    name: string,
    owner: string,
    phone?: string | null,
    rejectionReason?: string | null,
    rewards?:  {
      __typename: "ModelRewardConnection",
      nextToken?: string | null,
    } | null,
    serviceRadius?: number | null,
    state: string,
    updatedAt: string,
    verificationStatus?: BusinessVerificationStatus | null,
    website?: string | null,
    zipCode: string,
  } | null,
};

export type OnDeleteBusinessAnalyticsSubscriptionVariables = {
  filter?: ModelSubscriptionBusinessAnalyticsFilterInput | null,
  owner?: string | null,
};

export type OnDeleteBusinessAnalyticsSubscription = {
  onDeleteBusinessAnalytics?:  {
    __typename: "BusinessAnalytics",
    avgChildAge?: number | null,
    businessId: string,
    clicks?: number | null,
    createdAt: string,
    date: string,
    id: string,
    impressions?: number | null,
    owner?: string | null,
    redemptions?: number | null,
    revenue?: number | null,
    spend?: number | null,
    topRewardId?: string | null,
    updatedAt: string,
  } | null,
};

export type OnDeleteRewardSubscriptionVariables = {
  filter?: ModelSubscriptionRewardFilterInput | null,
  owner?: string | null,
};

export type OnDeleteRewardSubscription = {
  onDeleteReward?:  {
    __typename: "Reward",
    availableRadius?: number | null,
    business?:  {
      __typename: "Business",
      address: string,
      billingType?: BusinessBillingType | null,
      businessType?: BusinessBusinessType | null,
      city: string,
      cpmRate?: number | null,
      createdAt: string,
      currentSpend?: number | null,
      description?: string | null,
      email: string,
      id: string,
      isActive?: boolean | null,
      isVerified?: boolean | null,
      latitude?: number | null,
      logoKey?: string | null,
      logoUrl?: string | null,
      longitude?: number | null,
      monthlyBudget?: number | null,
      name: string,
      owner: string,
      phone?: string | null,
      rejectionReason?: string | null,
      serviceRadius?: number | null,
      state: string,
      updatedAt: string,
      verificationStatus?: BusinessVerificationStatus | null,
      website?: string | null,
      zipCode: string,
    } | null,
    businessId: string,
    category?: RewardCategory | null,
    clicks?: number | null,
    createdAt: string,
    currentRedemptions?: number | null,
    dailyLimit?: number | null,
    daysOfWeek?: Array< string | null > | null,
    description: string,
    endDate?: string | null,
    id: string,
    imageKey?: string | null,
    imageUrl?: string | null,
    impressions?: number | null,
    isActive?: boolean | null,
    owner?: string | null,
    pointCost: number,
    redemptionInstructions?: string | null,
    redemptionRecords?:  {
      __typename: "ModelRewardRedemptionConnection",
      nextToken?: string | null,
    } | null,
    redemptionType?: RewardRedemptionType | null,
    redemptions?: number | null,
    retailValue?: number | null,
    rewardType?: RewardRewardType | null,
    startDate: string,
    status?: RewardStatus | null,
    targetAgeRanges?: Array< string | null > | null,
    targetFamilySizes?: Array< number | null > | null,
    termsAndConditions?: string | null,
    title: string,
    totalAvailable?: number | null,
    updatedAt: string,
  } | null,
};

export type OnDeleteRewardRedemptionSubscriptionVariables = {
  filter?: ModelSubscriptionRewardRedemptionFilterInput | null,
  owner?: string | null,
};
// ...
export type OnUpdateBusinessSubscriptionVariables = {
  filter?: ModelSubscriptionBusinessFilterInput | null,
  owner?: string | null,
};

export type OnUpdateBusinessSubscription = {
  onUpdateBusiness?:  {
    __typename: "Business",
    address: string,
    billingType?: BusinessBillingType | null,
    businessType?: BusinessBusinessType | null,
    city: string,
    cpmRate?: number | null,
    createdAt: string,
    currentSpend?: number | null,
    description?: string | null,
    email: string,
    id: string,
    isActive?: boolean | null,
    isVerified?: boolean | null,
    latitude?: number | null,
    logoKey?: string | null,
    logoUrl?: string | null,
    longitude?: number | null,
    monthlyBudget?: number | null,
    name: string,
    owner: string,
    phone?: string | null,
    rejectionReason?: string | null,
    rewards?:  {
      __typename: "ModelRewardConnection",
      nextToken?: string | null,
    } | null,
    serviceRadius?: number | null,
    state: string,
    updatedAt: string,
    verificationStatus?: BusinessVerificationStatus | null,
    website?: string | null,
    zipCode: string,
  } | null,
};

export type OnUpdateBusinessAnalyticsSubscriptionVariables = {
  filter?: ModelSubscriptionBusinessAnalyticsFilterInput | null,
  owner?: string | null,
};
// ...
export type OnUpdateRewardSubscriptionVariables = {
  filter?: ModelSubscriptionRewardFilterInput | null,
  owner?: string | null,
};
// ...
export type OnUpdateRewardRedemptionSubscriptionVariables = {
  filter?: ModelSubscriptionRewardRedemptionFilterInput | null,
  owner?: string | null,
};